Replace NullScrobbleCache with simple MemoryScrobbleCache.
